fix: extract resolveBaseUrl to own file to avoid @git-glimpse/core build dep in tests
Importing from packages/action/src/index.ts in unit tests pulled in
@git-glimpse/core by package name, which requires a prior build step
that CI doesn't run for unit tests. Moving resolveBaseUrl to its own
file with a direct source-path type import fixes the resolution error.
